requirements.txt
networkx~=3.5
openai~=2.7.1
openai-whisper~=20250625
pydub~=0.25.1
transformers~=4.57.1
Flask~=3.1.2
redis~=7.0.1